feat: chatMemberUpdated 추가

+++ src/main/java/com/ebcho/engdabot/telegram/User.java
@@ -3,7 +3,7 @@ package com.ebcho.engdabot.telegram;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record User(
-	@JsonProperty("id") Long id,
+	@JsonProperty("id") Integer id,
 	@JsonProperty("is_bot") Boolean isBot,
 	@JsonProperty("first_name") String firstName,
 	@JsonProperty("username") String username,
